ms-retailer/pyproject.toml
[project]
name = "retailer"
version = "0.1.0"
description = "Kafka producer — simulates retail beer orders"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "lib-models",
    "confluent-kafka>=2.13.2",
    "opentelemetry-distro>=0.61b0",
    "opentelemetry-exporter-otlp>=1.40.0",
    "pydantic>=2.12.5",
]

[dependency-groups]
dev = [
    "ruff",
    "pyright",
    "pytest>=8",
]

[tool.uv.sources]
lib-models = { path = "../lib-models", editable = true }
